Disable non-FIPS code
Disable DSA. Two SSH specification constraints make it FIPS-incompatible:
- Fixed 1024-bit key size
- SHA-1 digest for signing/verification

Non-FIPS algorithms, or vendored implementations:
- UMAC
- ChaCha20-Poly1305
- Streamlined NTRU Prime 4591^761
- diffie-hellman-group1-sha1: Involves primes that are not
  FIPS-approved. Other DH variants use safe primes.
- Rijndael (AES)
- SHA-2 (256, 384, 512)

Implement FIPS-approved algorithms with OpenSSL v3:
- ED25519
- Curve25519

Ensure FIPS-incompatible algorithms are not advertised.

Tests:
- Skip disabled functionality
- RSA: Keys of size 1024 bits and below are incompatible with FIPS
  in signing operations. Use larger keys.

RSA subtlety: Depending on context ssh-rsa may refer to:
- Key type
- Public/host key signature algorithm

Per SSH specification, the ssh-rsa signature algorithm implies the
SHA-1 digest, which is not FIPS-compliant for signatures.
However, keys of type ssh-rsa may be used with rsa-ssh2-* algorithms.
In fact, keys of type ssh-rsa, rsa-sha2-256 and rsa-sha2-512 are
equivalent.
As it is desirable to be able to load keys of type ssh-rsa, we keep
it as a key type but do not advertise it as host key algorithm.