Disable non-FIPS code
Disable DSA. Two SSH specification constraints make it FIPS-incompatible:
- Fixed 1024-bit key size
- SHA-1 digest for signing/verification

Non-FIPS algorithms, or vendored implementations:
- UMAC
- ChaCha20-Poly1305
- Streamlined NTRU Prime 4591^761
- diffie-hellman-group1-sha1: Involves primes that are not
  FIPS-approved. Other DH variants use safe primes.
- Rijndael (AES)
- SHA-2 (256, 384, 512)
- Ed25519
- Curve25519

Ensure FIPS-incompatible algorithms are not advertised.

Tests:
- Skip disabled functionality
- RSA: Keys of size 1024 bits and below are incompatible with FIPS
  in signing operations. Use larger keys.
- Use ecdsa-sha2-nistp256 instead of ed25519 when FIPS mode is enabled.

RSA subtlety: Depending on context ssh-rsa may refer to:
- Key type
- Public/host key signature algorithm

Per SSH specification, the ssh-rsa signature algorithm implies the
SHA-1 digest, which is not FIPS-compliant for signatures.
However, keys of type ssh-rsa may be used with rsa-ssh2-* algorithms.
In fact, keys of type ssh-rsa, rsa-sha2-256 and rsa-sha2-512 are
equivalent (same representation).
As it is desirable to be able to load keys of type ssh-rsa, we keep
it as a key type but do not advertise it as host key algorithm.

ssh-rsa certificates: When `ssh -Q key` returns
`ssh-rsa-cert-v01@openssh.com`, it really means it supports:
- ssh-rsa-cert-v01@openssh.com
- rsa-sha2-256-cert-v01@openssh.com
- rsa-sha2-512-cert-v01@openssh.com

The first one uses non-FIPS-approved SHA-1 signatures, therefore only
the last two available in FIPS-mode.
While SHA-2 signatures are tolerated when verifying ssh-rsa-cert-v01
signatures, ssh uses SHA-1 when signing.
Tests that see ssh-rsa-cert-v01 in the `ssh -Q key` output attempt to
use all of them, which causes failures.
In addition, one may be concerned that, if we advertise
ssh-rsa-cert-v01, ssh peers may not know only some of them are
supported, causing confusing connection errors when they attempt to
authenticate.
For these reasons it was decided to disable support for all ssh-rsa
certificates in FIPS mode.

TODO:
- Module-Lattice-Based Key-Encapsulation Mechanism (ML KEM), apparently
  FIPS 203, when supported by OpenSSL FIPS provider (>= 3.5)

Upstream-Status: Pending